requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.11.0
filelock==3.12.2
fonttools==4.41.0
frozenlist==1.4.0
fsspec==2023.6.0
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
lightning-utilities==0.9.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.25.1
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
protobuf==4.23.4
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==2.0.5
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sympy==1.12
tensorboardX==2.6.1
torch==2.0.1
torchmetrics==1.0.1
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
yarl==1.9.2
